Resolve "Revise GeneralController Methods"

Closes #36

See merge request cs309/fall2021/1_cw_2!40

+++ Backend/FoodHub/src/main/java/foodhub/database/AdminRepository.java
@@ -11,6 +11,6 @@ public interface AdminRepository extends JpaRepository<Admin, Long> {
 	Admin findById(long id);
 	Admin findByUsername(String username);
 	
-	void deleteById(int id);
+	void deleteById(long id);
 	
 }
